# university-backend/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "University API",
        "description": "University Management System API Documentation",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://127.0.0.1:8000",
            "description": "Local API Server"
        }
    ],
    "paths": {
...
        "/api/submissions": {
            "post": {
                "tags": [
                    "Submissions"
                ],
                "summary": "Submit an assignment",
                "description": "SUBMIT ASSIGNMENT",
                "operationId": "86842154f1dffa008adb43a94e7c19a5",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "assignment_id",
                                    "student_id",
                                    "content"
                                ],
                                "properties": {
                                    "assignment_id": {
                                        "type": "integer",
                                        "example": 1
                                    },
                                    "student_id": {
                                        "type": "integer",
                                        "example": 1
                                    },
                                    "content": {
                                        "type": "string",
                                        "example": "https://github.com/my-solution"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Submission created successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "submission_id": {
                                            "type": "integer",
                                            "example": 1
                                        },
                                        "assignment_id": {
                                            "type": "integer",
                                            "example": 1
                                        },
                                        "student_id": {
                                            "type": "integer",
                                            "example": 1
                                        },
                                        "submitted_at": {
                                            "type": "string",
                                            "format": "date-time"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation Error"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/submissions/{submission_id}/grade": {
            "put": {
                "tags": [
                    "Submissions"
                ],
                "summary": "Grade a submission",
                "description": "GRADE SUBMISSION",
                "operationId": "241e62ed91218472804c4e5455537259",
                "parameters": [
                    {
                        "name": "submission_id",
                        "in": "path",
                        "description": "ID of the submission",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "score",
                                    "graded_by"
                                ],
                                "properties": {
                                    "score": {
                                        "type": "integer",
                                        "example": 85
                                    },
                                    "graded_by": {
                                        "type": "integer",
                                        "example": 2
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Graded successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "submission_id": {
                                            "type": "integer",
                                            "example": 1
                                        },
                                        "score": {
                                            "type": "integer",
                                            "example": 85
                                        },
                                        "graded_at": {
                                            "type": "string",
                                            "format": "date-time"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation Error"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        }
    },
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}